sparks: refactor
