+++ src/meson.build
@@ -564,7 +564,8 @@ common_sources = [
   'trx/game/shell/platform.c',
   'trx/game/sound/common.c',
   'trx/game/sound/ids.c',
-  'trx/game/sparks.c',
+  'trx/game/sparks/manager.c',
+  'trx/game/sparks/spawners.c',
   'trx/game/spawn.c',
   'trx/game/stats/common.c',
   'trx/game/stats/init.c',
