+++ src/trx/game/sparks/enum.h
@@ -1,0 +1,19 @@
+#pragma once
+
+enum {
+    // clang-format off
+    SPARK_F_NONE          = 0x0,
+    SPARK_F_SCALE         = 0x2,
+    SPARK_F_BLOOD         = 0x4,
+    SPARK_F_SPRITE        = 0x8,
+    SPARK_F_ROTATE        = 0x10,
+    SPARK_F_FX            = 0x40,
+    SPARK_F_ITEM          = 0x80,
+    SPARK_F_OUTSIDE       = 0x100,
+    SPARK_F_ALT_SPRITE    = 0x200,
+    SPARK_F_ATTACHED_POS  = 0x400,
+    SPARK_F_UNDERWATER    = 0x800,
+    SPARK_F_ATTACHED_NODE = 0x1000,
+    SPARK_F_GREEN         = 0x2000,
+    // clang-format on
+};
